7-Zip Version 23.01
- The page "Language" in 7-Zip's menu Tools/Options now shows
  information about selected translation, including the number of
  translated lines.

- 7-Zip now can use new ARM64 filter for compression to 7z and xz
  archives. ARM64 filter can increase compression ratio for data
  containing executable files compiled for ARM64 (AArch64)
  architecture.

- Also 7-Zip now parses executable files (that have exe and dll filename
  extensions) before compressing, and it selects appropriate filter for
  each parsed file:
  - BCJ or BCJ2 filter for x86 executable files,
  - ARM64 filter for ARM64 executable files.
  - Previous versions by default used x86 filter BCJ or BCJ2 for all
    exe/dll files.

- Default section size for BCJ2 filter was changed from 64 MiB to 240
  MiB. It can increase compression ratio for executable files larger
  than 64 MiB.

- UDF: support was improved.

- cpio: support for hard links.

- Some changes and optimizations in WIM creation code.

- When new 7-Zip creates multivolume archive, 7-Zip keeps in open state
  only volumes that still can be changed. Previous versions kept all
  volumes in open state until the end of the archive creation.

- 7-Zip for Linux and macOS now can reduce the number of simultaneously
  open files, when 7-Zip opens, extracts or creates multivolume archive.
  It allows to avoid the failures for cases with big number of volumes,
  bacause there is a limitation for number of open files allowed for a
  single program in Linux and macOS.

- There are optimizations in code for 7-Zip's context menu in Explorer:
  the speed of preparing of the menu showing was improved for cases when
  big number of files were selected by external program for context menu
  that contains 7-Zip menu commands.

- There are changes in code for the drag-and-drop operations to and from
  7-Zip File Manager.

- And the drag-and-drop operation with right button of mouse now is
  supported for some cases.

- The bugs were fixed:
  - ZIP archives: if multithreaded zip compression was performed with more
    than one file to stdout stream (-so switch), 7-zip didn't write
    "data descriptor" for some files.
  - ext4 archives: 7-Zip couldn't correctly extract symbolic link to
    directory from ext4 archives.
  - HFS and APFS archives: 7-Zip incorrectly decoded uncompressed blocks
    (64 KiB) in compressed forks.

+++ CPP/Windows/NationalTime.cpp
@@ -16,8 +16,8 @@ bool MyGetTimeFormat(LCID locale, DWORD flags, CONST SYSTEMTIME *time,
   if (numChars == 0)
     return false;
   numChars = ::GetTimeFormat(locale, flags, time, format,
-      resultString.GetBuf(numChars), numChars + 1);
-  resultString.ReleaseBuf_CalcLen(numChars);
+      resultString.GetBuf((unsigned)numChars), numChars + 1);
+  resultString.ReleaseBuf_CalcLen((unsigned)numChars);
   return (numChars != 0);
 }
 
@@ -29,8 +29,8 @@ bool MyGetDateFormat(LCID locale, DWORD flags, CONST SYSTEMTIME *time,
   if (numChars == 0)
     return false;
   numChars = ::GetDateFormat(locale, flags, time, format,
-      resultString.GetBuf(numChars), numChars + 1);
-  resultString.ReleaseBuf_CalcLen(numChars);
+      resultString.GetBuf((unsigned)numChars), numChars + 1);
+  resultString.ReleaseBuf_CalcLen((unsigned)numChars);
   return (numChars != 0);
 }
